+++ internal/sqlparser/statements.go
@@ -7,58 +7,6 @@ import (
 	"strings"
 )
 
-type Node any
-
-type Field struct {
-	Name      string  // "description"
-	DataType  Token   // "TEXT"
-	NotNull   bool    // true if NOT NULL, false if nullable
-	Default   *string // optional default value
-	IsPrimary bool    // true if PRIMARY KEY
-	IsUnique  bool    // true if UNIQUE
-}
-
-type TableType struct {
-	Name   string
-	Fields []Field
-}
-
-type EnumType struct {
-	Name   string
-	Values []string
-}
-
-type LogicalOp string
-
-const (
-	And     LogicalOp = "And"
-	Or      LogicalOp = "Or"
-	Illegal LogicalOp = ""
-)
-
-func toLogicOp(op string) LogicalOp {
-	switch op {
-	case "AND":
-		return And
-	case "OR":
-		return Or
-	default:
-		return Illegal
-	}
-}
-
-type ConditionOp string
-
-const (
-	EQUAL       ConditionOp = "="
-	NOTEQUAL    ConditionOp = "!="
-	LESSTHAN    ConditionOp = "<"
-	GREATERTHAN ConditionOp = ">"
-	BETWEEN     ConditionOp = "BETWEEN"
-	ISNULL      ConditionOp = "IS NULL"
-	NOTNULL     ConditionOp = "IS NOT NULL"
-)
-
 type Condition struct {
 	Left     []byte      // Column
 	Next     LogicalOp   // AND | OR | NOT
@@ -83,24 +31,6 @@ type InsertStatement struct {
 	InsertMode      []byte
 }
 
-type Ast struct {
-	l            *Lexer
-	Statements   []Node
-	currentToken Token
-	peekToken    Token
-}
-
-func NewAst(l *Lexer) *Ast {
-	return &Ast{
-		l: l,
-	}
-}
-
-func (a *Ast) NextToken() {
-	a.currentToken = a.peekToken
-	a.peekToken = a.l.NextToken()
-}
-
 func (a *Ast) Parse() error {
 	a.NextToken()
 	a.NextToken()
@@ -109,39 +39,10 @@ func (a *Ast) Parse() error {
 	case SELECT:
 		return a.parseSelect()
 	case INSERT:
-		return a.parserInsert()
+		return a.parseInsert()
 	default:
 		return fmt.Errorf("unexpected token: %s", a.currentToken.Literal)
 	}
-}
-
-func (a *Ast) ParseSchema() error {
-	a.NextToken()
-	a.NextToken()
-
-	for a.currentToken.Type != EOF {
-		switch a.currentToken.Type {
-		case CREATE:
-			a.NextToken()
-			if a.currentToken.Type == TABLE {
-				if err := a.parseTable(); err != nil {
-					return err
-				}
-				a.NextToken()
-			} else if a.currentToken.Type == TYPE {
-				if err := a.parseType(); err != nil {
-					return err
-				}
-				a.NextToken()
-			}
-		case SEMICOLON:
-			a.NextToken()
-		default:
-			return fmt.Errorf("unexpected token: %s", a.currentToken.Literal)
-		}
-	}
-
-	return nil
 }
 
 func (a *Ast) parseSelect() error {
@@ -250,7 +151,7 @@ func (a *Ast) parseSelect() error {
 	return nil
 }
 
-func (a *Ast) parserInsert() error {
+func (a *Ast) parseInsert() error {
 	stmt := &InsertStatement{}
 	a.NextToken()
 
@@ -296,124 +197,6 @@ func (a *Ast) parserInsert() error {
 				stmt.ReturningFields = append(stmt.ReturningFields, []byte(a.currentToken.Literal))
 			}
 		}
-	}
-
-	a.Statements = append(a.Statements, stmt)
-	return nil
-}
-
-func (a *Ast) parseTable() error {
-	stmt := &TableType{}
-
-	for a.currentToken.Type != STRING && a.currentToken.Type != EOF {
-		a.NextToken()
-	}
-
-	if a.currentToken.Type != STRING {
-		return fmt.Errorf("[PARSER_TABLE] unexpected token: %s wanted STRING", a.currentToken.Literal)
-	}
-	stmt.Name = a.currentToken.Literal
-	a.NextToken()
-
-	if a.currentToken.Type != LPAREN {
-		return fmt.Errorf("[PARSER_TABLE] unexpected token: %s wanted LPAREN", a.currentToken.Literal)
-	}
-	a.NextToken()
-
-	var fields []Field
-	// debug
-	idx := 0
-	for a.currentToken.Type != SEMICOLON && a.currentToken.Type != EOF {
-		var field Field
-		if a.currentToken.Type != STRING {
-			return fmt.Errorf("[PARSER_TABLE] unexpected token wanted columns name STRING got: %s field_idx: %d", a.currentToken.Literal, idx)
-		}
-		field.Name = a.currentToken.Literal
-		a.NextToken()
-
-		if a.currentToken.Type == IDENT && IsDatabaseType(a.currentToken.Literal) {
-			// TODO add datatype validation
-			field.DataType = a.currentToken
-			a.NextToken()
-		} else if a.currentToken.Type == STRING {
-			// ENUM type
-			field.DataType = Token{Type: ENUM, Literal: a.currentToken.Literal}
-			a.NextToken()
-		} else {
-			return fmt.Errorf("[PARSER_TABLE] expected datatype (IDENT or STRING), got: %s field_idx: %d", a.currentToken.Literal, idx)
-		}
-
-		for a.currentToken.Type != COMMA && a.currentToken.Type != RPAREN && a.currentToken.Type != EOF {
-			switch a.currentToken.Type {
-			case PRIMARY:
-				if a.peekToken.Type == KEY {
-					field.IsPrimary = true
-					a.NextToken() // consume PRIMARY
-					a.NextToken() // consume KEY
-				} else {
-					return fmt.Errorf("[PARSER_TABLE] expected KEY got: %s field_idx: %d", a.currentToken.Literal, idx)
-				}
-			case NOT:
-				if a.peekToken.Type == NULL {
-					field.NotNull = true
-					a.NextToken() // consume NOT
-					a.NextToken() // consume NULL
-				} else {
-					return fmt.Errorf("[PARSER_TABLE] expected NULL got: %s field_idx: %d", a.currentToken.Literal, idx)
-				}
-
-			case UNIQUE:
-				field.IsUnique = true
-				a.NextToken()
-			case DEFAULT:
-				a.NextToken()
-				if IsNowCompatible(field.DataType) && a.currentToken.Type == IDENT && a.peekToken.Type == LPAREN {
-					val := SqlNow(a.currentToken)
-					field.Default = &val
-					a.NextToken() // Consume now
-					a.NextToken() // Consume (
-					a.NextToken() // Consume )
-				} else if a.currentToken.Type == STRING || a.currentToken.Type == INT || a.currentToken.Type == TRUE || a.currentToken.Type == FALSE {
-					val := a.currentToken.Literal
-					field.Default = &val
-					a.NextToken()
-				} else {
-					return fmt.Errorf("[PARSER_TABLE] expected literal (STRING, INT, TRUE, FALSE, or now()), got: %s field_idx: %d", a.currentToken.Literal, idx)
-				}
-			default:
-				return fmt.Errorf("[PARSER_TABLE] unexpected token in field definition: %s field_idx: %d", a.currentToken.Literal, idx)
-			}
-		}
-		idx = idx + 1
-		fields = append(fields, field)
-		a.NextToken()
-	}
-
-	stmt.Fields = fields
-	a.Statements = append(a.Statements, stmt)
-	return nil
-}
-
-func (a *Ast) parseType() error {
-	stmt := &EnumType{}
-	a.NextToken()
-
-	if a.currentToken.Type != STRING {
-		return fmt.Errorf("unexpected token: %s WANTED STRING", a.currentToken.Literal)
-	}
-	stmt.Name = a.currentToken.Literal
-	a.NextToken()
-
-	if a.currentToken.Type != AS && a.peekToken.Type != ENUM {
-		return fmt.Errorf("failed parsing ENUM invalid got: %s", a.currentToken.Literal)
-	}
-	a.NextToken()
-
-	for a.currentToken.Type != RPAREN {
-		if a.currentToken.Type == STRING {
-			stmt.Values = append(stmt.Values, a.currentToken.Literal)
-		}
-		a.NextToken()
 	}
 
 	a.Statements = append(a.Statements, stmt)
@@ -545,57 +328,4 @@ func stringifyInsertStatement(stmt *InsertStatement) string {
 	}
 
 	return strings.TrimSpace(sb.String()) + ";"
-}
-
-func stringifyTableType(t *TableType) string {
-	var sb strings.Builder
-	sb.WriteString("CREATE TABLE IF NOT EXISTS ")
-	sb.WriteString(fmt.Sprintf("\"%s\"", t.Name))
-	sb.WriteString(" (\n")
-
-	for i, field := range t.Fields {
-		if field.DataType.Type == ENUM {
-			sb.WriteString(fmt.Sprintf("  \"%s\" \"%s\"", field.Name, field.DataType.Literal))
-		} else {
-			sb.WriteString(fmt.Sprintf("  \"%s\" %v", field.Name, field.DataType.Literal))
-		}
-
-		if field.NotNull {
-			sb.WriteString(" NOT NULL")
-		}
-		if field.IsUnique {
-			sb.WriteString(" UNIQUE")
-		}
-		if field.IsPrimary {
-			sb.WriteString(" PRIMARY KEY")
-		}
-		if field.Default != nil {
-			sb.WriteString(" DEFAULT ")
-			sb.WriteString(fmt.Sprintf("'%s'", *field.Default))
-		}
-
-		if i < len(t.Fields)-1 {
-			sb.WriteString(",\n")
-		} else {
-			sb.WriteString("\n")
-		}
-	}
-
-	sb.WriteString(");")
-	return sb.String()
-}
-
-func stringifyEnumType(e *EnumType) string {
-	var sb strings.Builder
-	sb.WriteString(fmt.Sprintf("CREATE TYPE \"%s\" AS ENUM (", e.Name))
-
-	for i, val := range e.Values {
-		sb.WriteString(fmt.Sprintf("'%s'", val))
-		if i < len(e.Values)-1 {
-			sb.WriteString(",")
-		}
-	}
-
-	sb.WriteString(");")
-	return sb.String()
 }
